+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added file activation support (for opening audiobooks from the file explorer and setting Audibly as the default app for opening audiobooks)
+
+### Fixed
+
+- Fixed a bug where:
+  - The app would crash when deleting the now playing audiobook from the library
+
 ## [2.0.8] - 04-24-2024
 
 ### Fixed